refactor: move click handler function of 近 into mapStore.js
